requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiomysql==0.2.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
frozenlist==1.4.0
idna==3.4
loging==0.0.1
magic-filter==1.0.12
multidict==6.0.4
pycparser==2.21
pydantic==1.10.13
PyMySQL==1.1.0
PyYAML==6.0.1
typing_extensions==4.8.0
yarl==1.9.2
